requirements.txt
playwright==1.58.0
python-dotenv==1.2.2
Requests==2.33.1
